fix: add exlintrc, fix all errors
