--- src/components/CardList/CardList.js
@@ -1,57 +1,0 @@
-import Card from "../Card/Card";
-import "./CardList.css";
-import taco from "../../images/recipes/taco.jpeg";
-
-function CardList({ title }) {
-  return (
-    <section className="card-list section">
-      <h3 className="card-list__title dark-color-text medium-text small-letter-spacing">
-        Категория: {title}
-      </h3>
-      <ul className="card-list__items">
-        <li>
-          <Card
-            cardTitle={"Тако с растительным фаршем"}
-            cardDescription={
-              "Быстрое и простое блюдо, из минимума ингредиентов, которое подойдет как для плотного обеда так и для сытного ужина"
-            }
-            to={"/recipe"}
-            image={taco}
-          />
-        </li>
-        <li>
-          <Card
-            cardTitle={"Тако с растительным фаршем"}
-            cardDescription={
-              "Быстрое и простое блюдо, из минимума ингредиентов, которое подойдет как для плотного обеда так и для сытного ужина"
-            }
-            to={"/recipe"}
-            image={taco}
-          />
-        </li>
-        <li>
-          <Card
-            cardTitle={"Тако с растительным фаршем"}
-            cardDescription={
-              "Быстрое и простое блюдо, из минимума ингредиентов, которое подойдет как для плотного обеда так и для сытного ужина"
-            }
-            to={"/recipe"}
-            image={taco}
-          />
-        </li>
-        <li>
-          <Card
-            cardTitle={"Тако с растительным фаршем"}
-            cardDescription={
-              "Быстрое и простое блюдо, из минимума ингредиентов, которое подойдет как для плотного обеда так и для сытного ужина"
-            }
-            to={"/recipe"}
-            image={taco}
-          />
-        </li>
-      </ul>
-    </section>
-  );
-}
-
-export default CardList;
