+++ src/components/Contacts/Contacts.css
@@ -79,14 +79,12 @@
 }
 
 .contacts-label {
-  padding-bottom: 7px;
-}
-
-.contacts-label::after {
-  content: "*";
+  display: flex;
+  flex-direction: column;
 }
 
 .contacts-input {
+  margin-top: 7px;
   color: var(--color1);
   border: 1px solid rgba(61, 61, 64, 0.15);
   padding: 10px 20px;
