+++ src/components/Footer/Footer.jsx
@@ -1,10 +1,12 @@
-import "./Footer.css";
+import React from 'react';
+import './Footer.css';
 
 function Footer() {
   return (
     <footer className="footer">
       <h5 className="footer__copy small-text dark-color-text">
-        &copy; {new Date().getFullYear()}
+        &copy;
+        {new Date().getFullYear()}
       </h5>
     </footer>
   );
