+++ src/components/NavButtons/NavButtons.jsx
@@ -1,14 +1,19 @@
-import { useNavigate } from "react-router-dom";
-import "./NavButtons.css";
-import back from "../../images/icons/back-left.svg";
-import top from "../../images/icons/top.svg";
+import React from 'react';
+import { useNavigate } from 'react-router-dom';
+import './NavButtons.css';
+import back from '../../images/icons/back-left.svg';
+import top from '../../images/icons/top.svg';
 
 function NavButtons() {
-  let navigate = useNavigate();
+  const navigate = useNavigate();
 
   return (
     <div className="nav-buttons">
-      <button onClick={() => navigate(-1)} className="nav-buttons__item">
+      <button
+        type="button"
+        className="nav-buttons__item"
+        onClick={() => navigate(-1)}
+      >
         <img src={back} alt="Назад" title="Вернуться назад" />
       </button>
       <a href="#top">
