+++ src/components/PageNotFound/PageNotFound.jsx
@@ -1,9 +1,10 @@
-import { useNavigate } from "react-router-dom";
-import "./PageNotFound.css";
-import back from "../../images/icons/back.svg";
+import React from 'react';
+import { useNavigate } from 'react-router-dom';
+import './PageNotFound.css';
+import back from '../../images/icons/back.svg';
 
 function PageNotFound() {
-  let navigate = useNavigate();
+  const navigate = useNavigate();
 
   return (
     <section className="page-not-found section">
@@ -13,7 +14,11 @@ function PageNotFound() {
       <h3 className="page-not-found__description dark-color-text medium-text small-letter-spacing">
         Возвращайтесь скорее назад
       </h3>
-      <button className="page-not-found__button" onClick={() => navigate(-1)}>
+      <button
+        type="button"
+        className="page-not-found__button"
+        onClick={() => navigate(-1)}
+      >
         <img
           className="page-not-found__button-image"
           src={back}
